FAN-11 funcionalidad consultar lista

+++ src/main/resources/schema.sql
@@ -56,7 +56,7 @@ CREATE TABLE IF NOT EXISTS `products` (
   `codeproduct` INT NOT NULL AUTO_INCREMENT,
   `nameproduct` VARCHAR(100) NOT NULL,
   `categoryproduct` VARCHAR(45) NOT NULL,
-  `priceproduct` INT NOT NULL,
+  `priceproduct` FLOAT NOT NULL,
   `currencyproduct` VARCHAR(5) NOT NULL,
   `stockproduct` INT NOT NULL,
   `statusproduct` VARCHAR(45) NOT NULL,
